refactor backend var
