--- src/components/CreateProfile/backend.js
@@ -1,3 +1,0 @@
-const backend = 'https://mental-healthy-backend.herokuapp.com';
-
-export default backend;
